+++ ContentPatcher/Framework/ContentPatcherAPI.cs
@@ -131,7 +131,7 @@ namespace ContentPatcher.Framework
             // format name
             name = $"{mod.UniqueID}{InternalConstants.ModTokenSeparator}{name}";
             this.Monitor.Log($"{mod.Name} added a token: {name}", LogLevel.Trace);
-            ModProvidedToken token = new ModProvidedToken(name, mod, valueProvider);
+            ModProvidedToken token = new ModProvidedToken(name, mod, valueProvider, this.Monitor);
 
             // add token
             this.AddModToken(token);